- changed deltaPhi to DeltaPhi - moved all functions dependent on hcal rechits to HcalRechitHelper.cxx - defined JetPassesHWQual function; replaced duplicate code - added Pass_ branches to both event level and perjet trees - added jetIndex_DepthTagCand and jetIndex_InclTagCand global variables; added to both  event level and perjet trees - added pass preselection function that actually defines jetIndex_DepthTagCand and jetIndex_InclTagCand global variables
