+++ DisplacedHcalJetAnalyzer/DisplacedHcalJetAnalyzer/DisplacedHcalJetAnalyzer.h
@@ -103,7 +103,6 @@ public :
    map<string,string>   tree_output_vars_string;
    map<string,vector<float>>  tree_output_vars_vec;
 
-
    vector<string> jet_treenames;
    map<string,TTree*>   jet_tree_output;
    map<string,bool>     jet_tree_output_vars_bool;  
@@ -132,6 +131,9 @@ public :
    float Muon_PhiVectorSum = -9999.9;
 
    int N_PFJets_ToSave = 4;
+
+   int jetIndex_DepthTagCand = -1;
+   int jetIndex_InclTagCand = -1;
 
    // ----- Variables ----- //
 
@@ -763,27 +765,29 @@ public :
    virtual void   ProcessEvent( Long64_t jentry );
    // TriggerHelper.cxx
    virtual void   SetTriggerNames( string infilepath, string hist_name );
-   // Object Helper.cxx
+   // ObjectHelper.cxx
    virtual float  DeltaR( float eta1, float eta2, float phi1, float phi2);
-   virtual double deltaPhi( double phi1, double phi2);
-   // virtual double deltaR( double eta1, double phi1, double eta2, double phi2);
+   virtual double DeltaPhi( double phi1, double phi2);
+   virtual bool   JetPassesHWQual( int jet_index, float &deltaR );
+   virtual bool   IsMuonIsolatedTight( int muon_index ); 
+   virtual float  GetElectronEffectiveAreaMean( int ele_index );
+   virtual bool   IsElectronIsolatedTight( int ele_index );
+   virtual float  TransverseLeptonMass( float pT, float phi );
+   virtual float  PhiVectorSum( float pT, float phi );
+   // HcalRechitHelper.cxx
    virtual vector<int>             GetRechitMult( int idx_llp, float deltaR_cut );
-   virtual vector<vector<float>>   GetEnergyProfile( int idx_llp, float deltaR_cut );
    virtual vector<float>           GetMatchedHcalRechits_Jet( int idx_jet, float deltaR_cut );
    virtual vector<vector<float>>   GetHcalRechitValues_Jet( int idx_jet );
+   virtual vector<vector<float>>   GetEnergyProfile( int idx_llp, float deltaR_cut );
    virtual vector<float>           GetEnergyProfile_Jet( int idx_jet, float deltaR_cut );
    virtual vector<pair<float,int>> Get3RechitE_Jet( int idx_jet, float deltaR_cut );
-   virtual vector<float> GetEtaPhiSpread_Jet( int idx_jet, float deltaR_cut );
-   virtual int           GetTimingTowers_Jet( int idx_jet, float deltaR_cut );
-   virtual int           GetDepthTowers_Jet( int idx_jet, float deltaR_cut );
-   virtual int           GetDepthTowers_Jet_lowE( int idx_jet, float deltaR_cut );
-   virtual int           GetTotalTowers_Jet( int idx_jet, float deltaR_cut );
-   virtual vector<float> GetTDCavg_Jet( int idx_jet, float deltaR_cut );
-   virtual bool  IsMuonIsolatedTight( int muon_index ); 
-   virtual float GetElectronEffectiveAreaMean( int ele_index );
-   virtual bool  IsElectronIsolatedTight( int ele_index );
-   virtual float TransverseLeptonMass( float pT, float phi );
-   virtual float PhiVectorSum( float pT, float phi );
+   virtual vector<float>           GetEtaPhiSpread_Jet( int idx_jet, float deltaR_cut );
+   virtual int                     GetTimingTowers_Jet( int idx_jet, float deltaR_cut );
+   virtual int                     GetDepthTowers_Jet( int idx_jet, float deltaR_cut );
+   virtual int                     GetDepthTowers_Jet_lowE( int idx_jet, float deltaR_cut );
+   virtual int                     GetTotalTowers_Jet( int idx_jet, float deltaR_cut );
+   virtual vector<float>           GetTDCavg_Jet( int idx_jet, float deltaR_cut );
+
    // TruthInfoHelper.cxx
    virtual void   SetLLPVariables();
    virtual bool   isRechitValid(float RechitEnergy, int RechitDepth);
@@ -801,8 +805,9 @@ public :
    // EventHelper.cxx
    virtual float  GetEventRuntime( clock_t clock_start, Long64_t init_entry, Long64_t current_entry );
    virtual void   ResetGlobalEventVars();
-   virtual int    PassDisplacedJetHLT();
-   virtual bool   PassEventPreselection( int passedHLT = -1 );
+   virtual bool   PassL1SingleLLPJet();
+   virtual bool   PassHLTDisplacedJet();
+   virtual bool   PassEventPreselection( bool passedHLT = false );
    virtual bool   PassWPlusJetsSelection();
    virtual bool   PassLeptonVeto();
    virtual bool   PassZmumuSelection();
@@ -818,8 +823,8 @@ public :
    virtual void   DeclareOutputJetTrees();
    virtual void   ResetOutputBranches( string treename );
    virtual vector<pair<float,float>> TrackMatcher( int jetIndex, vector<uint> jet_track_index );
-   virtual void   FillOutputTrees( string treename );
-   virtual void   FillOutputJetTrees( string treename, int jetIndex );
+   virtual void   FillOutputTrees( string treename, map<string, bool> Pass_EventSelections = {} );
+   virtual void   FillOutputJetTrees( string treename, int jetIndex, map<string, bool> Pass_EventSelections = {} );
    virtual void   WriteOutputTrees();
    virtual void   SetHistCategories();
    virtual void   BookHists();
